+++ NEWS.md
@@ -1,3 +1,8 @@
+# comunicacion 0.0.0.9007
+
+* Agrego funciones `lbl_int()`, `lbl_decimal()`, `lbl_percent()` y `lbl_money()` para formatear números con separador "." para miles y "," para decimales, además de "%" para porcentajes y "$" para moneda.
+* Fix formats integer y decimals en `gt_theme_dnmye()`.
+
 # comunicacion 0.0.0.9006
 
 * Agrega función `crear_presentacion()` con template {xaringan}